# services/cell-api/internal/api/openapi_gen.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "description": "Personal access token (con_pat_…) or service-account token (con_sa_…).",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "Sluicio-Session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Sluicio cell-api. Authenticate with a session cookie (browser) or a Bearer token — a personal access token or a service-account token (see Settings → Service accounts). Generated from the route table; do not edit by hand.",
    "title": "Sluicio API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/integrations/{id}/completion-firings/{iid}/handle": {
      "post": {
        "operationId": "post_api_v1_integrations_id_completion_firings_iid_handle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "iid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create integration",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/v1/me/password": {
      "post": {
        "operationId": "post_api_v1_me_password",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "summary": "Create me",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
...
}